player数据保存代码重构外加血量保存

+++ Assets/Scripts/ScriptableObj/So_PlayerProperties.cs
@@ -5,10 +5,11 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "so_SceneRouteList", menuName = "Scriptable Objects/PlayerProperties")]
 public class So_PlayerProperties : ScriptableObject 
 {
+  
     public Vector3 position=Vector3.zero;
 
     public Quaternion roatition=Quaternion.identity;
     public bool HaveGun;
 
-    public float HealthValue;
+    public int HealthValue;
 }
